requirements.txt
tornado==6.1
opencv-python==4.5.1.48
bs4==0.0.1
numpy==1.20.2
Pillow==7.0.0
pytesseract==0.3.7